requirements.txt
streamlit
pandas
numpy
folium
streamlit-folium
xarray
netCDF4
psycopg2-binary
google-api-python-client
google-auth-httplib2
google-auth-oauthlib